requirements.txt
fastapi==0.108.0
uvicorn==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.6.2
python-multipart
